Move types into formatfile
